Move SSL initialisation
