+++ src/main/scala/net/ripe/rpki/publicationserver/Boot.scala
@@ -1,25 +1,13 @@
 package net.ripe.rpki.publicationserver
 
-import java.io.FileInputStream
-import java.security.KeyStore
-
 import akka.actor.{ActorRef, ActorSystem, OneForOneStrategy, SupervisorStrategy}
-import akka.http.scaladsl.Http
 import akka.http.scaladsl.server.Directives._
-import akka.stream.ActorMaterializer
 import akka.util.Timeout
 import com.softwaremill.macwire._
-import com.softwaremill.macwire.akkasupport._
-import com.typesafe.sslconfig.akka.AkkaSSLConfig
-import akka.http.scaladsl.{ConnectionContext, HttpsConnectionContext, Http}
-import javax.net.ssl._
-import net.ripe.logging.SysStreamsLogger
+import akka.http.scaladsl.{ConnectionContext, Http, HttpsConnectionContext}
 import net.ripe.rpki.publicationserver.store.XodusDB
 import net.ripe.rpki.publicationserver.metrics._
-import org.slf4j.{LoggerFactory, Logger}
-import akka.pattern.ask
-
-import io.prometheus.client.exporter.common.TextFormat
+import org.slf4j.{Logger, LoggerFactory}
 import io.prometheus.client._
 
 import scala.concurrent.duration._
@@ -27,23 +15,40 @@ import scala.concurrent.Await
 import scala.concurrent.Future
 import akka.http.scaladsl.Http.ServerBinding
 import java.{util => ju}
-import java.io.IOException
-import scala.util.{Failure,Success,Try}
+
+import net.ripe.rpki.publicationserver.util.SSLHelper
+
+import scala.util.{Failure, Success, Try}
 
 
 
 object Boot extends App {
+  lazy val conf = wire[AppConfig]
+  lazy val logger = setupLogging(conf)
+
+  override final def main(args: Array[String]) = {
+    val sslHelper = new SSLHelper(conf, logger)
+
+    val https: HttpsConnectionContext = {
+      sslHelper.connectionContext match {
+        case Success(v) => v
+        case Failure(e) =>
+          logger.error("Error while creating SSL context, exiting", e)
+          // EX_DATAERR (65) The input data was incorrect in some way.
+          sys.exit(65)
+      }
+    }
+
+    new PublicationServerApp(conf, https, logger).run()
+  }
+
   def setupLogging(conf: AppConfig) = {
     System.setProperty("LOG_FILE", conf.locationLogfile)
     LoggerFactory.getLogger(this.getClass)
   }
-
-  lazy val conf = wire[AppConfig]
-  lazy val logger = setupLogging(conf)
-  new PublicationServerApp(conf, logger).run()
 }
 
-class PublicationServerApp(conf: AppConfig, logger: Logger) extends RRDPService {
+class PublicationServerApp(conf: AppConfig, https: ConnectionContext, logger: Logger) extends RRDPService {
     
   implicit val system = ActorSystem.create(Math.abs(new ju.Random().nextLong()).toString())
   implicit val dispatcher = system.dispatcher
@@ -68,17 +73,6 @@ class PublicationServerApp(conf: AppConfig, logger: Logger) extends RRDPService 
 
     val publicationService = new PublicationService(conf, stateActor)
 
-    val https: HttpsConnectionContext = {
-      val connectionContext = Try(ConnectionContext.https(sslContext()));
-      connectionContext match {
-        case Success(v) => v
-        case Failure(e) =>
-          logger.error("Error while creating SSL context, exiting", e)
-          // EX_DATAERR (65) The input data was incorrect in some way.
-          sys.exit(65)
-      }
-    }
-
     this.httpsBinding = Http().bindAndHandle(
       publicationService.publicationRoutes,
       interface = conf.serverAddress,
@@ -102,72 +96,4 @@ class PublicationServerApp(conf: AppConfig, logger: Logger) extends RRDPService 
                 .terminate(hardDeadline = 3.seconds))
             .flatMap(_ => system.terminate)
   }
-
-  private def sslContext(): SSLContext = {
-    val sslContext = SSLContext.getInstance("TLS")
-    sslContext.init(getKeyManagers, getTrustManagers, null)
-    sslContext
-  }  
-
-  private def getTrustManagers: Array[TrustManager] = {
-    if (conf.publicationServerTrustStoreLocation.isEmpty) {
-      logger.info(
-        "publication.server.truststore.location is not set, skipping truststore init"
-      )
-      null
-    } else {
-      val trustStore = KeyStore.getInstance("JKS")
-      val tsPassword: Array[Char] =
-        if (conf.publicationServerTrustStorePassword.isEmpty) {
-          null
-        } else {
-          conf.publicationServerTrustStorePassword.toCharArray
-        }
-      logger.info(
-        s"Loading HTTPS certificate from ${conf.publicationServerTrustStoreLocation}"
-      )
-      trustStore.load(
-        new FileInputStream(conf.publicationServerTrustStoreLocation),
-        tsPassword
-      )
-      val tmf = TrustManagerFactory.getInstance("SunX509")
-      tmf.init(trustStore)
-      tmf.getTrustManagers
-    }
-  }
-
-  private def getKeyManagers: Array[KeyManager] = {
-    if (conf.publicationServerKeyStoreLocation.isEmpty) {
-      if (conf.getConfig.getBoolean(
-            "publication.spray.can.server.ssl-encryption"
-          )) {
-        logger.error(
-          "publication.spray.can.server.ssl-encryption is ON, but publication.server.keystore.location " +
-            "is not defined. THIS WILL NOT WORK!"
-        )
-      } else {
-        logger.info(
-          "publication.server.keystore.location is not set, skipping keystore init"
-        )
-      }
-      null
-    } else {
-      val ksPassword = if (conf.publicationServerKeyStorePassword.isEmpty) {
-        null
-      } else {
-        conf.publicationServerKeyStorePassword.toCharArray
-      }
-      val keyStore = KeyStore.getInstance("JKS")
-      logger.info(
-        s"Loading HTTPS certificate from ${conf.publicationServerKeyStoreLocation}"
-      )
-      keyStore.load(
-        new FileInputStream(conf.publicationServerKeyStoreLocation),
-        ksPassword
-      )
-      val kmf = KeyManagerFactory.getInstance("SunX509")
-      kmf.init(keyStore, ksPassword)
-      kmf.getKeyManagers
-    }
-  } 
 }
