+++ src/test/scala/net/ripe/rpki/publicationserver/integration/PublicationIntegrationTest.scala
@@ -6,16 +6,15 @@ import akka.testkit.{TestActorRef, TestKit}
 import net.ripe.rpki.publicationserver.Binaries.{Base64, Bytes}
 import net.ripe.rpki.publicationserver.store.ObjectStore
 import net.ripe.rpki.publicationserver.store.fs.RsyncRepositoryWriter
-import net.ripe.rpki.publicationserver.PublicationServerBaseTest
-import net.ripe.rpki.publicationserver.AppConfig
-import net.ripe.rpki.publicationserver.Hashing
+import net.ripe.rpki.publicationserver.{AppConfig, Hashing, PublicationServerApp, PublicationServerBaseTest}
 import org.slf4j.LoggerFactory
-import net.ripe.rpki.publicationserver.PublicationServerApp
 import java.nio.file._
 import java.net.URI
 
+import akka.http.scaladsl.{ConnectionContext, Http, HttpsConnectionContext}
 import akka.http.scaladsl.model.MediaType
 import akka.http.scaladsl.model.HttpCharsets
+import net.ripe.rpki.publicationserver.util.SSLHelper
 
 class PublicationIntegrationTest
     extends PublicationServerBaseTest
@@ -48,7 +47,7 @@ class PublicationIntegrationTest
       override lazy val publicationServerTrustStorePassword = "123456"
       override lazy val publicationServerKeyStorePassword = "123456"
     }
-    server = new PublicationServerApp(conf, logger)
+    server = new PublicationServerApp(conf, new SSLHelper(conf, logger).connectionContext.get, logger)
     server.run()
     client = new PublicationServerClient()
   }
